perubahan int ke float

+++ PrakFungsiSwitch/PrakFungsiSwitch.cpp
@@ -26,7 +26,7 @@ int perkalian(int a, int b)
     return a * b;
 }
 
-float pembagian(int a, int b)
+float pembagian(float a, float b)
 {
     return a / b;
 }
